desarrollo hu consultas

+++ src/main/java/com/BobElAlquilador/demo/repository/ConversacionRepository.java
@@ -1,11 +1,20 @@
 package com.BobElAlquilador.demo.repository;
 
-
 import com.BobElAlquilador.demo.model.Conversacion;
+import com.BobElAlquilador.demo.model.Persona;
+
+import org.antlr.v4.runtime.atn.SemanticContext.AND;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ConversacionRepository extends JpaRepository<Conversacion,Integer> {
-    //definir más metodos de ser necesarios
+public interface ConversacionRepository extends JpaRepository<Conversacion, Long> {
+
+  //List<Conversacion> findByCliente(Persona cliente);
+  //List<Conversacion> findByEmpleado(Persona empleado);
+  
+
 }
